+++ app/src/main/java/com/picketing/www/business/domain/reservation/Reservation.java
@@ -1,5 +1,10 @@
-package com.picketing.www.business.domain.show.reservation;
+package com.picketing.www.business.domain.reservation;
 
+import java.util.ArrayList;
+import java.util.List;
+
+import com.picketing.www.business.domain.User;
+import com.picketing.www.business.domain.show.Show;
 import com.picketing.www.persistence.table.BaseEntity;
 
 import jakarta.persistence.Entity;
@@ -9,6 +14,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -17,20 +23,26 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
+@Table(name = "RESERVATION")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
 @Builder(access = AccessLevel.PROTECTED)
-@Table(name = "RESERVED_SEAT")
 @Getter(AccessLevel.PROTECTED)
-public class ReservedSeat extends BaseEntity {
+public class Reservation extends BaseEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 
 	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "reservation_id")
-	private Reservation reservation;
+	@JoinColumn(name = "show_id")
+	private Show show;
 
-	// private TimeScheduleSeatGrade timeScheduleSeatGrade;
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "user_id")
+	private User user;
+
+	@OneToMany(mappedBy = "reservation")
+	private List<ReservedSeat> reservedSeatList = new ArrayList<>();
+
 }
