--- app/src/main/java/com/picketing/www/business/domain/show/seat/SeatFactory.java
@@ -1,16 +1,0 @@
-package com.picketing.www.business.domain.show.seat;
-
-import org.springframework.stereotype.Component;
-
-import com.picketing.www.presentation.dto.request.seat.SaveSeatRequest;
-
-@Component
-public class SeatFactory {
-
-	public Seat createSeat(SaveSeatRequest saveSeatRequest) {
-		return Seat.builder()
-			.show(saveSeatRequest.show())
-			.timeScheduleId(saveSeatRequest.timeScheduleId())
-			.build();
-	}
-}
