--- app/src/main/java/com/picketing/www/persistence/repository/show/seat/SeatRepository.java
@@ -1,8 +1,0 @@
-package com.picketing.www.persistence.repository.show.seat;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface SeatRepository extends JpaRepository<Seat, Long> {
-}
